T-011: PR-2: Create prop-driven assignees.tsx in v2 + update legacy
