+++ x-pack/solutions/security/plugins/security_solution/public/flyout/attack_details/components/assignees.tsx
@@ -5,161 +5,21 @@
  * 2.0.
  */
 
-import type { FC } from 'react';
-import React, { memo, useCallback, useMemo, useState } from 'react';
-
-import {
-  EuiButtonIcon,
-  EuiContextMenu,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiPopover,
-  EuiPopoverTitle,
-  EuiToolTip,
-  useGeneratedHtmlId,
-} from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
-import { isNonLocalIndexName } from '@kbn/es-query';
-import { getEmptyTagValue } from '../../../common/components/empty_value';
-import { UsersAvatarsPanel } from '../../../common/components/user_profiles/users_avatars_panel';
-import { useBulkGetUserProfiles } from '../../../common/components/user_profiles/use_bulk_get_user_profiles';
-import { useLicense } from '../../../common/hooks/use_license';
-import { useUpsellingMessage } from '../../../common/hooks/use_upselling';
-import { useAttackAssigneesContextMenuItems } from '../../../detections/hooks/attacks/bulk_actions/context_menu_items/use_attack_assignees_context_menu_items';
-import { useAttacksPrivileges } from '../../../detections/hooks/attacks/bulk_actions/use_attacks_privileges';
-import { useInvalidateFindAttackDiscoveries } from '../../../attack_discovery/pages/use_find_attack_discoveries';
+import React, { memo, useMemo } from 'react';
+import type { EsHitRecord } from '@kbn/discover-utils';
+import { buildDataTableRecord } from '@kbn/discover-utils';
+import { Assignees as AssigneesV2 } from '../../../flyout_v2/attack/main/components/assignees';
 import { useAttackDetailsContext } from '../context';
-import { useHeaderData } from '../hooks/use_header_data';
-import { HEADER_ASSIGNEES_ADD_BUTTON_TEST_ID } from '../constants/test_ids';
-
-const AssigneesButton: FC<{
-  isDisabled: boolean;
-  toolTipMessage: string;
-  onClick: () => void;
-}> = memo(({ onClick, isDisabled, toolTipMessage }) => (
-  <EuiToolTip position="bottom" content={toolTipMessage}>
-    <EuiButtonIcon
-      aria-label={i18n.translate(
-        'xpack.securitySolution.attackDetailsFlyout.header.assignees.ariaLabel',
-        { defaultMessage: 'Update assignees' }
-      )}
-      data-test-subj={HEADER_ASSIGNEES_ADD_BUTTON_TEST_ID}
-      iconType="plusCircle"
-      onClick={onClick}
-      isDisabled={isDisabled}
-    />
-  </EuiToolTip>
-));
-AssigneesButton.displayName = 'AssigneesButton';
 
 /**
- * Assignees block for the Attack details flyout header.
- * Follows the same pattern as status_popover_button: useAttackDetailsContext + useHeaderData
- * + useAttackAssigneesContextMenuItems, with EuiPopover + EuiContextMenu.
+ * Context bridge: wraps the v2 Assignees component for the legacy attack details flyout.
+ * Reads searchHit and refetch from context, builds a DataTableRecord, then delegates to v2.
  */
 export const Assignees = memo(() => {
-  const { attackId, indexName, refetch } = useAttackDetailsContext();
-  const isRemoteDocument = useMemo(() => isNonLocalIndexName(indexName), [indexName]);
-  const { alertIds, assignees } = useHeaderData();
-  const invalidateFindAttackDiscoveries = useInvalidateFindAttackDiscoveries();
-  const { hasIndexWrite, hasAttackIndexWrite, loading: privilegesLoading } = useAttacksPrivileges();
-  const isPlatinumPlus = useLicense().isPlatinumPlus();
-  const upsellingMessage = useUpsellingMessage('alert_assignments');
+  const { searchHit, refetch } = useAttackDetailsContext();
+  const hit = useMemo(() => buildDataTableRecord(searchHit as EsHitRecord), [searchHit]);
 
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-  const togglePopover = useCallback(() => setIsPopoverOpen((prev) => !prev), []);
-  const closePopover = useCallback(() => setIsPopoverOpen(false), []);
-  const assigneesTitleId = useGeneratedHtmlId();
-
-  const attacksWithAssignees = useMemo(
-    () => [{ attackId, relatedAlertIds: alertIds, assignees }],
-    [attackId, alertIds, assignees]
-  );
-
-  const onSuccess = useCallback(() => {
-    refetch();
-    invalidateFindAttackDiscoveries();
-  }, [refetch, invalidateFindAttackDiscoveries]);
-
-  const { items, panels } = useAttackAssigneesContextMenuItems({
-    attacksWithAssignees,
-    closePopover,
-    onSuccess,
-    telemetrySource: 'attacks_page_flyout_header',
-  });
-
-  const uids = useMemo(() => new Set(assignees), [assignees]);
-  const { data: assignedUsers } = useBulkGetUserProfiles({ uids });
-
-  const hasPermission =
-    Boolean(hasIndexWrite) && Boolean(hasAttackIndexWrite) && isPlatinumPlus && !privilegesLoading;
-
-  const toolTipMessage =
-    upsellingMessage ??
-    i18n.translate('xpack.securitySolution.attackDetailsFlyout.header.assignees.popoverTooltip', {
-      defaultMessage: 'Assign attack',
-    });
-
-  const button = useMemo(
-    () => (
-      <AssigneesButton
-        onClick={togglePopover}
-        isDisabled={isRemoteDocument}
-        toolTipMessage={toolTipMessage}
-      />
-    ),
-    [togglePopover, toolTipMessage, isRemoteDocument]
-  );
-
-  if (!hasPermission) {
-    return (
-      <div data-test-subj="attack-details-flyout-header-assignees-empty">{getEmptyTagValue()}</div>
-    );
-  }
-
-  return (
-    <EuiFlexGroup
-      gutterSize="none"
-      responsive={false}
-      data-test-subj="attack-details-flyout-header-assignees"
-    >
-      {assignedUsers && assignedUsers.length > 0 && (
-        <EuiFlexItem grow={false}>
-          <UsersAvatarsPanel userProfiles={assignedUsers} maxVisibleAvatars={2} />
-        </EuiFlexItem>
-      )}
-      <EuiFlexItem grow={false}>
-        <EuiPopover
-          aria-labelledby={assigneesTitleId}
-          button={button}
-          isOpen={isPopoverOpen}
-          closePopover={closePopover}
-          panelPaddingSize="none"
-          data-test-subj="attack-details-flyout-header-assignees-popover"
-        >
-          <EuiPopoverTitle id={assigneesTitleId} paddingSize="m">
-            {i18n.translate(
-              'xpack.securitySolution.attackDetailsFlyout.header.assignees.popoverTitle',
-              {
-                defaultMessage: 'Manage assignees',
-              }
-            )}
-          </EuiPopoverTitle>
-          <EuiContextMenu
-            panels={[
-              {
-                id: 0,
-                items,
-              },
-              ...panels,
-            ]}
-            initialPanelId={0}
-            data-test-subj="attack-details-flyout-header-assignees-context-menu"
-          />
-        </EuiPopover>
-      </EuiFlexItem>
-    </EuiFlexGroup>
-  );
+  return <AssigneesV2 hit={hit} onAttackUpdated={refetch} />;
 });
 
 Assignees.displayName = 'Assignees';
